refactor: mypage refactor
