+++ src/app/mypage/setting_and_private/page.tsx
@@ -1,4 +1,4 @@
-import MenuList from '../MenuList';
+import MenuList from './MenuList';
 
 function Mypage() {
   return (
